requirements.txt
bctpy==0.6.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.0.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
MarkupSafe==2.1.2
networkx==3.0
nibabel==5.0.1
nilearn==0.10.0
numpy==1.23.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psutil==5.9.4
pyg-lib==0.1.0+pt113cu116
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
torch==1.13.1+cu116
torch-cluster==1.6.0+pt113cu116
torch-geometric==2.1.0
torch-scatter==2.1.0+pt113cu116
torch-sparse==0.6.16+pt113cu116
torch-spline-conv==1.2.1+pt113cu116
torchaudio==0.13.1+cu116
torchmetrics==0.11.1
torchvision==0.14.1+cu116
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
